+++ example/src/demos/MultiRender.tsx
