+++ example/src/demos/Pointcloud.tsx
