+++ example/vite.config.ts
@@ -6,8 +6,6 @@ export default defineConfig({
   resolve: {
     alias: [
       { find: 'react', replacement: path.resolve('./node_modules/react') },
-      //{ find: '@react-spring/three', replacement: path.resolve('./node_modules/@react-spring/three/index.js') },
-      { find: 'react-three-fiber', replacement: path.resolve('../packages/fiber/src/web/index.tsx') },
       { find: '@react-three/fiber', replacement: path.resolve('../packages/fiber/src/web/index.tsx') },
     ],
   },
